Reorder stock files into diff folders
