+++ backend/src/main/java/com/is442g1t4/gpa/stock/Stock.java
@@ -1,4 +1,4 @@
-package com.is442g1t4.gpa.stock.model;
+package com.is442g1t4.gpa.stock;
 
 import java.util.Date;
