+++ backend/src/main/java/com/is442g1t4/gpa/stock/StockService.java
@@ -1,11 +1,7 @@
-package com.is442g1t4.gpa.stock.service;
+package com.is442g1t4.gpa.stock;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import com.is442g1t4.gpa.stock.exception.StockNotFoundException;
-import com.is442g1t4.gpa.stock.model.Stock;
-import com.is442g1t4.gpa.stock.repository.StockRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,11 +28,10 @@ public class StockService {
 
     public Stock addStock(Stock stock) {
         Optional<Stock> retrievedStock = getStockByTicker(stock.getStockTicker());
-        if (retrievedStock.isPresent()){
+        if (retrievedStock.isPresent()) {
             return null;
         }
         return stockRepository.save(stock);
     }
-
 
 }
