+++ backend/src/main/java/com/is442g1t4/gpa/stock/enums/Country.java
@@ -1,0 +1,5 @@
+package com.is442g1t4.gpa.stock.enums;
+
+public enum Country {
+
+}
