+++ backend/src/main/java/com/is442g1t4/gpa/stock/enums/Industry.java
@@ -1,0 +1,5 @@
+package com.is442g1t4.gpa.stock.enums;
+
+public enum Industry {
+
+}
