+++ backend/src/main/java/com/is442g1t4/gpa/stock/exceptions/StockNotFoundException.java
@@ -1,4 +1,4 @@
-package com.is442g1t4.gpa.stock.exception;
+package com.is442g1t4.gpa.stock.exceptions;
 
 public class StockNotFoundException extends RuntimeException {
     public StockNotFoundException(String message) {
