--- backend/src/main/java/com/is442g1t4/gpa/stock/model/JsonbDateFormat.java
@@ -1,5 +1,0 @@
-package com.is442g1t4.gpa.stock.model;
-
-public @interface JsonbDateFormat {
-
-}
