+++ backend/src/main/java/com/is442g1t4/gpa/stock/stockPrice/StockPriceRepository.java
@@ -1,4 +1,4 @@
-package com.is442g1t4.gpa.stock.repository;
+package com.is442g1t4.gpa.stock.stockPrice;
 
 import java.util.Date;
 import java.util.Optional;
@@ -7,8 +7,6 @@ import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.stereotype.Repository;
-
-import com.is442g1t4.gpa.stock.model.StockPrice;
 
 @Repository
 public interface StockPriceRepository extends MongoRepository<StockPrice, ObjectId> {
